Add validation.StructList for better alternative RuleSet syntax
